% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namaKU.R
\name{namaKU}
\alias{namaKU}
\title{Making acronym and sequence number from long name}
\usage{
namaKU(df, location, vocal = TRUE, number = 3, seq = TRUE)
}
\arguments{
\item{df}{data frame that contains the long name}

\item{location}{the name of column that contain long name}

\item{vocal}{\code{TRUE} to keep vocal letter, \code{FALSE} to remove vocal letter}

\item{number}{number of letter to keep from each word}

\item{seq}{\code{TRUE} to add sequence number, \code{FALSE} to not add sequence number}
}
\value{
a data frame containing the acronym name based on the parameters
}
\description{
Making acronym and sequence number from long name
}
\examples{
# Load necessary packages
library(dplyr)
library(rlang)
# Assuming 'df' is your data frame and it has a column 'AreaName'
d <- data.frame(id = seq(1:3), AreaName = c("loNDON", "Port Alegre", 'MakaSSar'), sub = c('a', 'b', 'c'))
# Generate acronyms with sequence
d <- namaKU(df = d, location = AreaName, vocal = FALSE, number = 3, seq = TRUE)
# Check the results
d
}
